requirements.txt
streamlit
tensorflow
pillow
h5py
numpy
pandas
scikit-learn